requirements.txt
aiogram>=2.25.1
aiogTTS>=1.1.1
gpytranslate>=1.5.1
openai>=0.27.7
tabulate>=0.9.0
Pillow>=9.5.0